Explicitly setting background to dark
Tmux is having one of his moments again.
Vim's colors are off in Tmux, but fine outside.
Setting the background to dark explicitly, and making sure that tmux
uses the right terminfo (By setting env's TERM variable to either
screen-256color or xterm-256color) does the trick to fix it.